requirements.txt
PyQt5
matplotlib
python-dateutil